Integrate sage (#11)
I want to use the (python based) computer algebra system sage (sagemath) starting from chapter 7.

My intention is to use sage to support the solutions of chapter 7. If this works well (it is an experiment) I will continue with sage in the following chapters.

In this MR I changed the infrastructure to work with sage directly instead of pure python (CI/CD workflow). I made a proof of concept in chapter 7 demonstrating the technical feasibility.